pyproject.toml
[project]
name = "gambabot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "discord-py>=2.3.0",
    "python-dotenv>=1.0.0",
]
